+++ packages/firestore/lite/index.d.ts
@@ -46,7 +46,7 @@ export function setLogLevel(logLevel: LogLevel): void;
 export interface FirestoreDataConverter<T> {
   toFirestore(modelObject: T): DocumentData;
   toFirestore(modelObject: Partial<T>, options: SetOptions): DocumentData;
-  fromFirestore(snapshot: QueryDocumentSnapshot): T;
+  fromFirestore(snapshot: QueryDocumentSnapshot<DocumentData>): T;
 }
 
 export class FirebaseFirestore {
