+++ packages/firestore/lite/src/api/transaction.ts
@@ -95,7 +95,11 @@ export class Transaction implements firestore.Transaction {
     options?: firestore.SetOptions
   ): Transaction {
     const ref = validateReference(documentRef, this._firestore);
-    const convertedValue = applyFirestoreDataConverter(ref._converter, value);
+    const convertedValue = applyFirestoreDataConverter(
+      ref._converter,
+      value,
+      options
+    );
     const parsed = this._dataReader.parseSetData(
       'Transaction.set',
       ref._key,
